Create components for PrivateRoute and AuthButton, Add type annotations
